requirements.txt
requests==2.28.1
pandas==1.3.5
retrying==1.3.3
pyspark==3.3.0
delta-spark==2.0.0